+++ System.Private.CoreLib/System/Linq/AnyAll.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections.Generic;
+
+namespace System.Linq
+{
+    public static partial class Enumerable
+    {
+        public static bool Any<T>(this IEnumerable<T> values)
+        {
+            var enumerator = values.GetEnumerator();
+            return enumerator.MoveNext();
+        }
+    }
+}
